+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/contentful/contentful.php/compare/1.1.0...HEAD)
+## [1.2.0](https://github.com/contentful/contentful.php/tree/1.2.0) (2017-05-16)
 
 ### Added
 * Implemented `ResourceArray::getItems` to allow access to the values of a `ResourceArray` as an actual PHP array.